Suricata update working
