+++ src/electron/windows/zui-window.ts
@@ -67,9 +67,10 @@ export abstract class ZuiWindow {
 
   load() {
     this.beforeLoad()
-    const url = env.isDevelopment
-      ? `http://localhost:3000${this.path}?id=${this.id}&name=${this.name}`
-      : `file://${this.path}.html?id=${this.id}&name=${this.name}`
+    const url =
+      env.isDevelopment && !env.isTest
+        ? `http://localhost:3000${this.path}?id=${this.id}&name=${this.name}`
+        : `file://${this.path}.html?id=${this.id}&name=${this.name}`
     return this.ref.loadURL(url)
   }
