--- src/plugins/brimcap/create-cli.ts
@@ -1,9 +1,0 @@
-import {env} from "src/zui"
-import BrimcapCLI from "./brimcap-cli"
-import path from "path"
-
-export function createCli() {
-  const commandName = env.isWindows ? "brimcap.exe" : "brimcap"
-  const bin = path.join(env.zdepsPath, commandName)
-  return new BrimcapCLI(bin)
-}
