+++ src/plugins/brimcap/packets/is-pcap.ts
