+++ src/plugins/brimcap/packets/types.ts
@@ -1,0 +1,2 @@
+export type DownloadArgs = [pool: string, uid: string]
+export const DOWNLOAD = "brimcap.downloadPackets"
