+++ src/plugins/brimcap/suricata/util.ts
