+++ src/plugins/brimcap/zeek/queries.ts
