+++ src/ppl/detail/RelatedConns.tsx
@@ -14,7 +14,7 @@ import firstLast from "./util/firstLast"
 import formatDur from "./util/formatDur"
 import {useSelector} from "react-redux"
 import Results from "src/js/state/Results"
-import {SURICATA_CONNS} from "src/plugins/suricata/ids"
+import {SURICATA_CONNS} from "src/plugins/brimcap/suricata/ids"
 import time from "src/js/models/time"
 
 const id = SURICATA_CONNS
